Moved class ModelCheckpointExt
